pyproject.toml
[tool.poetry]
name = "cursofast"
version = "0.1.0"
description = ""
authors = ["NicholasDRR <nicholasreis00@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
pymongo = "^4.5.0"
python-dotenv = "^1.0.0"
pillow = "^10.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
